Vector per-element comparisons
These comparisons may be used to return a vector of booleans, which contains the result of each individual comparison.

These comparisons may be performed with Rhs vectors or scalars. If it is a vector, respective elements will be compared; if it is a scalar, all lhs elements will be compared with the scalar.