New tests | VectorClamp
VectorClamp is potentially underperforming, as it can be beaten by VectorClampMin followed by VectorClampMax.

However, this only happens in release mode after the first load of iterations of tests. VectorClamp is always faster than the combination on the first time around.

Additionally, VectorClamp is always faster in debug mode. As such, it is likely that the performance difference is due to some unlikely-to-occur-in-the-wild situation that the compiler is optimising the full program for - after all, it's unlikely one would ever be doing these millions of exact same operations across multiple vectors at a frequent rate.