Reference vectors updated
Previously, vectors of references could only be constructed from individual arguments; they could not reference vectors as single arguments due to how things are handled.

We now have a new EmuCore reference wrapper, which allows deferred provision of a reference. This is a very barebones reference and has little safety checks, but is intended to be used identically to std::reference_wrapper as far as general interactions go, with the exception of allowing default construction to leave provision of a reference until later.

## Notes regarding this new reference wrapper
- This is not intended for use outside of class internals
- This is specifically for scenarios such as vector copy-construction, where we initialise the array and then set the references individually immediately after initialisation (before the constructor may fully end).
- Where possible, std::reference_wrapper should be used instead.
- Comments highlight this

These null-starting references are the new reference wrappers contained within an EmuMath vector.