Some general additions all around
This was a result of tests of bitwise operations with floats, which had not yet been implemented outside of SIMD and thus caused a compiler error.

This has been amended, although bitwise |, &, and ^ operations with floating points remain undefined due to increased complexity compared to the existing specialisations as of this commit.

Also added some specialisations for NoOverflowT to make use of the floating point specialisations should it contain a floating point. Note that this is only the case for bitwise operations as fiddling with only known bits guarantees no overflows. Other functors, such as arithmetic, remain unspecialised as using the specialisation indirectly would result in an implicit loss of no-overflow safety, which ruins the entire point of NoOverflowT.

FastVector4f now has bitwise operation functions, including an implementation for NOT which had previously not been implemented as an operator.
   - This operation uses the new index mask functions to get an __m128 vector with all bits set to 1, and than returns a vector built with `andnot(data_, mask_)`

Both normal and fast vectors now have vectorwise min/max/minmax functions, which return a vector containing the lowest or greatest elements of respective indices within 2 vectors, or within 1 vector and a rhs scalar.

FastVector4f now has arithmetic named member functions.