Constexpr sin and cos functors
Made for the sake of being able to easily form constexpr rotation matrices.

Calculation uses a taylor series algorithm of X iterations, where X is a custom number that may be provided as a template argument. This X defaults to 3, which provides an approximate value at a reasonable speed.

Tan and arc- variants are not present.