Perspective projection matrix generation
I still have no idea what the hell happened in the previous, discarded iteration, but do_tan_constexpr now works correctly and is not producing erroneous values under certain circumstances.

Nevertheless, we now have a perspective matrix which can be automatically calculated with a frustum based on the Y-angle of FOV, aspect ratio, and the near and far planes.
Alternatively, instead of FOV angle and aspect ratio, one can provided custom left-, right-, bottom-, and top-edges of the frustum. Near and far planes are still required in this version.

The output matrix is equivalent to:
{ (2*near)/(right-left), 0, (right+left)/(right-left). 0 }
{ 0, (2*near)/(top-bottom), (top+bottom)/(top-bottom), 0 }
{ 0, 0, -((far+near)/(far-near)), -((2*far*near)/(far-near)) }
{ 0, 0, -1, 0 }