Matrices may now be made with contiguous scalar arguments
This comes with a `MatrixMake` function, which can be considered a Matrix form of `std::make_tuple`, although it requires a small bit extra: the arguments to create the matrix type instantiation.

This works regardless of the contiguous order of a matrix, and will adapt to the major order so that data is taken with contiguous semantics.

For example, here are the results of two identical calls but for major order:
```
EmuMath::Helpers::MatrixMake<2, 2, float, true>(1, 2, 3, 4);
Resulting Matrix:
| 1 3 |
| 2 4 |
```

```
EmuMath::Helpers::MatrixMake<2, 2, float, false>(1, 2, 3, 4);
Result Matrix:
| 1 2 |
| 3 4 |
```

These results may be confirmed in this commit via the end of the starting prints before tests are triggered.