Matrix arithmetic
Not including all bits yet, but using matrix multiplication test as a checkpoint.

Note: tests show `EmuMath` matrices are faster than `DirectXMath` matrices for multiplication in release, but slower in debug.
They can be faster in debug if we do a dump of data and broadcast the dumped data, but this results in slower performance in release. It could be reasonable to do a debug switch to use optimised debug behaviour, however this would complicate the code unnecessarily.