Colour validation layer template | Clamped and Wrapped colour aliases
`ClampedColour` and `WrappedColour` both ensure that contained colours remain in a valid state by either clamping or wrapping them (respectively) into the valid range.

Both are aliases of the `_colour_validation_layer` template, which is designed to not be used directly but instead through aliases such as these.

`_colour_validation_layer` provides a mostly matching interface to `Colour`, meeting a contract for what is expectable for template functions.