Issue correction and optimisation for fast perlin 3D gen
So we weren't actually shuffling the final row for forming dot products; we just did 2 pointless shuffles.

This has been corrected.

You may be thinking "oh no, the fast noise table generation just slowed down!", but you would be thinking wrong! With this commit also comes an optimisation which makes use of the final, unused slot of our gradient registers for a duplicate Z.
This duplicate slot is used to perform blends instead of shuffles to get our intermediate registers for shuffling. These blends are then used to perform an 'optimised shuffle' (that is, it calls our SIMD shuffle helper which has an optimised specialisation to achieve the same effect).

This optimisation makes up for a slight issue with shuffle throughput, and takes advantage of functioning on multiple ports unlike shuffles. This leads our new execution being marginally (negligibly so, in fact) faster than previously.

So, with that said, this small optimisation has thankfully made it so that our correction to function correctly is not harming performance!