Dead info removed | Sets removed
Vectors can no longer be `Set` after construction. Originally this was the same as a `Copy` for value-containing Vectors, but reference-containing Vectors would update their references on a call to `Set`.

This behaviour has been removed to follow C++'s focus on non-nullable references. Now references within Vectors will focus on the same memory from birth until death, and no side effects will be modifying that unless an explicit type is used to do so.
   - This also means that dangling references are not modifiable, so now reference Vectors have a clear (and intended) lifetime compared to value Vectors.