Validity tests...kind of
Visual studio SFINAE support is abysmal.

We cannot use `std::enable_if_t` to toggle `operator=` availability as visual studio seems to draw false negatives for `operator=(const this_type&)`.
A probable alternative for this is to do some general tests within the type itself, and add a partial specialisation to `std::is_assignable`, although this certainly is not preferable.