Matrix mutation seemingly complete + speed cmp
This resulted in a weird adventure into disassembly when the zero-cost abstraction approach seemed to be failing.

As it turns out, a by-hand implementation was using `float, float` args, but the abstraction approach was using `float, long double` args. This somehow went under the radar for long enough for me to decide I had to check the assembly instructions, and noticed double and conversion instructions being used.
I have never felt so embarrassed by something that only I saw, so I decided to share it here too.