STL Iterators for EmuMath::Vector
Provides STL random_access_iterator support to `EmuMath::Vector`, allowing use of the following familiar functions:
- `begin`
- `end`
- `cbegin`
- `cend`
- `rbegin`
- `rend`
- `crbegin`
- `crend`

This provides the same `value_type` interface, and does not expose underlying reference storage, allowing it to work the same way with both value and reference Vectors as the general interface with said Vectors already worked.

The naming convention of the STL-like functions betrays the usual Emu naming convention to allow `EmuMath::Vector` to be plugged directly into any algorithms that make use of this STL syntax.