Vector Fmsub | Rename to Fma to Fmadd
Fmsub is actually emulated via fmadd, since we only have `std::fma` and, for the sake of future support should it be made constexpr, I have chosen to *not* write it in inline assembly (since constexpr math functions are far more likely than constexpr asm blocks).

With the above said, it should also be noted that `std::fma` is not constexpr evaluable. As such, when at least 1 argument is floating-point, `EmuCore::do_fmadd` (and by extension, `EmuCore::do_fmsub`) is not constexpr evaluable, which by extension also means this is the case for Vector (and future Matrix) fmadd/fmsub calls.
   - In other words, it is never constexpr-evaluable in situations where it makes a difference.