Conversion construction
A minor, but not especially safe, issue has become clear with areas where visual studio's false-positive warnings are silenced.
Although we never access the same index twice when allowing moves, we could potentially be moving from a reference Vector/Matrix which, although unlikely, *could* contain multiple references to the same object, which results in actually repeated-using a moved-from object.

Nothing is directly planned to fix this, and the niche, lacking likelihood pushes it down to a very low priority. Nonetheless, an issue will be raised regarding this.