Matrix STL Iterators
Due to the contiguity change, STL iterators have been implemented for Matrix.

The implementation is purely the same iterator as the underlying Vector storage - since we store in a single flattened, contiguous EmuMath Vector, that's all we need to do. This also meets the prior issue's suggestion for this access to be contiguous, although that has become standardised for all flattened index accesses in Matrix.

This was not a part of the issue requesting this contiguity change, but was a quick addition and sticks to the theme of the update.