Quick correction to exception-throwing (if invalid) access functions
We were using `ostringstream` to build exception messages when constexpr dynamic allocation was available, but `ostringstream` is not valid for use in this way. This meant that these functions could not be used in a constexpr context on C++20 and above.

Has been updated to use progressive string appending after a reserve similar to `Vector`'s implementation.