Reintroduction of tests | Flatten function for matrices
After some personal debate regarding whether contiguous access should be provided directly to a Matrix, it has been decided to keep Matrix's underlying data as private as it is not necessarily guaranteed to be an EmuMath Vector (although is unlikely to be changed due to the consistency assistance it now provides).

Instead, a `Flatten` member function is now available, which does allow output of references, but as an EmuMath Vector containing references, rather than a direct reference to the underlying Vector itself.

This `Flatten` function defaults to copy output as per all non-`at` functions, but is still classed as an `Access` function under the `#pragma region` categories.

Additionally, it allows a further level of abstraction to be applied to this flattening when desired, as a Matrix may be flattened as either column-major or row-major regardless of its actual major order. This does, however, default to a Matrix's actual major-order. This means that:
   - `matrix.Flatten<true>()` is predictable, and will always output a Vector version of the Matrix in column-major.
   - `matrix.Flatten<false>()` is predictable, and will always output a Vector version of the Matrix in row-major.
   - `matrix.Flatten()` is not predictable if the major-order of the Matrix is not known. However, there is the guarantee that this will always be the same contiguous order of the Matrix itself, and can be predicted via `matrix.is_column_major`