Some general prep for FastVector conversions
This has led to a huge overhaul of EmuSIMD's underlying implementation, as it became clear that a further layer of indirection would be needed to remain reasonable maintainable.

This extra layer is most likely to be optimised out, but will give better ability to update available intrinsics at a later date (which relates back to a previous commit commenting on my lack of ARM intrinsic knowledge).

This approach will also allow the ability to add a scalar emulation when no intrinsics are available, so we can stop code from shitting itself if changing to a platform without intrinsics.