C++20 | Adds alternative Rect assignments | Allows shorthand construction of Vectors via variadic
In this case, shorthand describes being able to use:
`EmuMath::Vector<4, float> vec = { 1, 2, 3, 4 };`
instead of:
`EmuMath::Vector<4, float> vec  = EmuMath::vector<4, float>(1, 2, 3, 4);`

This additionally allows syntax such as:
`vec = { 5, 6, 7, 8 };`

## Extras resulting from this
Where using a `Vector<1, T_>`, this allows the constructor to further be used with simple scalars, for example:
`EmuMath::Vector<1, float> vec = 7;`
Note that this does not allow uses such as `EmuMath::Vector<4, float> vec = 7;`.

## Start of std upgrade
This is primarily possible due to the ease of use and maintainability of the C++20 feature `explicit(bool)`, which allows us to conditionally specific a constructor as `explicit` without need for SFINAE mechanisms.

This is the first benefit being taken advantage of from a decision to finally transition this library to C++20, as the next official standard update is not far away (relatively speaking).

Note that, despite this upgrade, current content is not intended to be upgraded to C++20 just yet. This is more so to allow new content as of now to take advantage of the 2020 standard's features, to reduce workload when a full makeover is given.

A time for a full upgrade is undecided; a stronger analysis of more intense projects that this library is planned for will be required first.