Discard search for speed issue
Based on disassembly, it appears that the speed difference is a result of the templatised approach alongside an additional layer of abstraction (as we're using `FastVector`s to represent majors).

Additionally, the difference between this and `DirectX`'s Matrix is generally ~0.002-0.004 microseconds (in terms of per-iteration mean). This can stack up significantly, but right now is not considered a significant-enough bottleneck to focus on.